# Remove isAlphaNum in LBS decodeLenient